updated top level comments
at this point, the code has been modified to the point that it is no longer soley r0gueSch0lar's work. I now understand that this is describing a fairly simple geometric shape which is the union of rectangular prisms, difference of cylinders, and intersection of a large cylinder. it can be expressed in mathematical notation with about 15-20 symbols, so I don't think anyone can claim to "own" this shape.

I have also simplified the code quite a bit. it is much shorter. there were many unnecessary factors of 1/2 and spacing that just made it hard to read.